Use Vec<u32> in Command immediate data

+++ wgpu-core/src/command/compute_command.rs
@@ -26,7 +26,7 @@ pub enum ComputeCommand<R: ReferenceType> {
         offset: u32,
 
         /// The immediate data to be written.
-        data: Vec<u8>,
+        data: Vec<u32>,
     },
 
     DispatchWorkgroups([u32; 3]),
